Clean up and rearrange packages
- Remove TLS configurability
- Move packages that can be easily moved and are internal to /internal
- Update some package names to be more representative of their function
- Remove obsolete/unused code
