--- grpc-server/testdata/bundle.pem
@@ -1,32 +1,0 @@
-TEST CERTIFICATE ONLY - NOT FOR REAL USAGE 
------BEGIN CERTIFICATE-----
-MIIFTzCCAzegAwIBAgIUBviL+dMgIkMf7R8nPspjyb+LcGgwDQYJKoZIhvcNAQEL
-BQAwNzENMAsGA1UECgwEQ05DRjETMBEGA1UECwwKVGlua2VyYmVsbDERMA8GA1UE
-AwwIdGVzdC1rZXkwHhcNMjIwMzE4MTUyMjIyWhcNMzIwMzE1MTUyMjIyWjA3MQ0w
-CwYDVQQKDARDTkNGMRMwEQYDVQQLDApUaW5rZXJiZWxsMREwDwYDVQQDDAh0ZXN0
-LWtleTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBALonX7dDury4h569
-ZQlBSUTL5+gdFxKMw0rrMwVxcHDm8Q4t9OlomHZu37Epc91XRfui4YEYglTo3Zzl
-36gwkAPcfvGQ0KUtC1nRXs0FNOfvIm+Uv++l8S2B3BVaEABvcdpMQPkyZnKLhssd
-0jF7/SFH96fdY8Xmwf0LPHcobjrQGURLU8S7mtLMJUiC0ByDTfTye6cTLFdELqzu
-3LYbnT8ZceLOvLDmERJoQEgJxg2Lhr1H2TMmEUDy+svAEEyEUqYIXrPhrfN2wWUY
-rIZgUupDOxvpSIzTeo/kHNrI6tbJ3laaz93Xb/0c6v5Wq/B7pViEQr5gPw+Q/h2+
-3padZIwuKAOCZLdILl0gQbUSYJ3CrOTzXBp8k++FLA8Oa8ENG3JzFXZpfsPV4Ovb
-T/dulrG3zHXLjEZ/H7oNyvOLV92fJDBguGJEUeRqD8p6bmhcctygGQBBiqaKyxnl
-+mGadsDrMWyegm7ua25hHnFgLofzBUEAiXGh1z0XdZAizT/oyPRn2TGAslgedoqK
-wcmvOeZp+GVZi9qE8Fbd/cbdcbMWHoVBafSUh3nKiMAWaD8VeB8NKA3p2PK9j0Qn
-L73kCyovDI4n9pUkkNv3pWId5DXKDAUyVgC3/ZeIwMj8lKufoG97lUyf4gOJj3xC
-6fyZxaZqNk6vrGPJdY8vkF+r4e9NAgMBAAGjUzBRMB0GA1UdDgQWBBRnYKNBhu9S
-IEqFqyKeJ8R7hGqagzAfBgNVHSMEGDAWgBRnYKNBhu9SIEqFqyKeJ8R7hGqagzAP
-BgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUAA4ICAQBybdkX6LoOsHKFBSRE
-85yCc/tq5DNA24BAyvLBnHa4oenVt4g+cDT+eZKe/rZlHZwH6CSuWBI3XG0eyxjt
-1TNkGtPcPZFHt/xe0x1nFE1XuWtWF9pwzJvJ3fDVd3IwyXoTbQ040jaNC3csVsMV
-yLJNPKDT4ULVM76L89Gh/GsuWlJM9ocwfAWmfWTIUnDtJWVDw+TKvli4+IFPVStE
-dFqB95AybJ1pI+0OSUioLkvrd+7udFTxn6QaqKGPDOocA+cGkR+oHmV1g72ucPaF
-iTeTIT+/rzpGmmdoi1SGNdm+8+0afRHTRCjc14wRXSBDgOLtIk8f6R4lo3ZZY9xO
-2PM6h1khWThjt6aEEVWL40dS9we6iIbTlc/auYR97EWMIZnlaHjOg1A4tn/XIlVR
-bJ6M1x1eqCQ0bv/VLte671FXuUIgbvu6XIIDv3kW+/YDWKNkiIm2uGvrR/wFMz7r
-xqyH0PzVVO9C21nNd09ZFRZM5+SF8N9NavSL2Q7m1RV8E8Boj1kJOAjYmeWYbDFB
-hq5CZKXw59+WKga0ETZb3CHA7SXK0S8+8lhHdusXsV4vdgB6jL/hDH6UbFi7cxB9
-mU5Z9/wAcNUR8+UCdiW+ZpK76VFqEgWIZJzhswZ4mXET6nqypYw5XE0s3WUz2+Sh
-HFFHJvKS/BrOklUc9AUGFKujZw==
------END CERTIFICATE-----
