--- grpc-server/testdata/server-key.pem
@@ -1,53 +1,0 @@
-TEST PRIVATE KEY ONLY - NOT FOR REAL USAGE 
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQC6J1+3Q7q8uIee
-vWUJQUlEy+foHRcSjMNK6zMFcXBw5vEOLfTpaJh2bt+xKXPdV0X7ouGBGIJU6N2c
-5d+oMJAD3H7xkNClLQtZ0V7NBTTn7yJvlL/vpfEtgdwVWhAAb3HaTED5MmZyi4bL
-HdIxe/0hR/en3WPF5sH9Czx3KG460BlES1PEu5rSzCVIgtAcg0308nunEyxXRC6s
-7ty2G50/GXHizryw5hESaEBICcYNi4a9R9kzJhFA8vrLwBBMhFKmCF6z4a3zdsFl
-GKyGYFLqQzsb6UiM03qP5BzayOrWyd5Wms/d12/9HOr+Vqvwe6VYhEK+YD8PkP4d
-vt6WnWSMLigDgmS3SC5dIEG1EmCdwqzk81wafJPvhSwPDmvBDRtycxV2aX7D1eDr
-20/3bpaxt8x1y4xGfx+6Dcrzi1fdnyQwYLhiRFHkag/Kem5oXHLcoBkAQYqmissZ
-5fphmnbA6zFsnoJu7mtuYR5xYC6H8wVBAIlxodc9F3WQIs0/6Mj0Z9kxgLJYHnaK
-isHJrznmafhlWYvahPBW3f3G3XGzFh6FQWn0lId5yojAFmg/FXgfDSgN6djyvY9E
-Jy+95AsqLwyOJ/aVJJDb96ViHeQ1ygwFMlYAt/2XiMDI/JSrn6Bve5VMn+IDiY98
-Qun8mcWmajZOr6xjyXWPL5Bfq+HvTQIDAQABAoICAQCHPeXXIji/tRyqohSOdcUC
-W1W/l6rUijmz605lDPZQwCevUooVLS1fFcwkTOZlj2tDlyFYBfNiNtASlhs4eReY
-BpCfdcNvzVrqxSansrmuK1kMUbhkJl4i3q6DQKxRdKX1n+KwaQJvA5lJZf/4fYj4
-re1qInjDJZQYABrMwy3aQqeoq0VPr7Cap0AK/yatIGP5qlVVm6NiPyHd96eYElXa
-quTHj8Uci/kpM15IN/mQi6a3S0SsWWK9mgnFD2OIA7Z786bB7xrOv59fkF5/Penf
-UjrWW1SfI2Fuup6QANpEc+K5br3IASXWcTT01QkTrPECbwyCfTAzYfaMq8fRCsYp
-ppzla+F/YOKT2xtNyBj3dYnLyqEoapiL1knzoPlXAfnwu5Nb/YdQrt/AnX25Cga7
-8l3Rk/nMaGtOwv9Wj7HrRMC9a754+ILFWyGYQw8GTww4a4nJ7aXrXQJTh8WYGFRp
-hQUbOX3lWT4BVhm2ENQxaqGsB/sDSSMke9dhf7OvBnxtArxDQvFHulWkPQd5idYD
-pDFwlgCLCp9USeQp3cwieS1ukMPYVHjx+pgGknUqzEPscjCmqgttOhMRS2f1m2KK
-bs4kJBzipwM981qx4WTp2rhXBKbFmSQYR5Rv8vgjzPaaCxg96YH+V+/4Jd5fQtw3
-R9mYa1Xg43FLuiR/brkbIQKCAQEA86qFknjZKIe/5CCwaa5lGXj80fjdJqX1is2v
-YJmrw/Kupt7dTS8+6ZQPHj74eGab0gyww5OVWKCsvKWgTdq+qIL9eQBl1v9B7Ckn
-aRX6SfMDMbftlzQQYy2redz/X6PJaUggOsHWtNWP2gbKFeeBaZbDOjDCV0xlsThC
-9gIp2Yhya1EDcMXod6/FTuEE4oyMwktU6bLxwKv9fu0Oqo7g1Wr9COM4xal1sLhq
-FUPpguYhWkKBR8sF/LoOuf18stdcMv/X/194Ry5Bc1TVb9oIDV6E7M8Wamfc5ngh
-w5PwRyWqkziBhTOgUgJ9n+frBBV9ieNxPOtnsjxfbyOkXiIs6QKCAQEAw5OYZaTu
-g9yXlI70sYVKcF1OMZZ+yY+MgRaDy2if/0Mr2ZGGh4UhBfSXM1oCKYJt+Oxx3v/F
-YbSzGDeJ0ZefcOU5RL8jePx3/rn8pSjfCTYGZG0DE3bFdCzbWRdmcgz3RiFHc12V
-eKmrIES5rk3Jv78SzrvbMOStU7DNMX82PlVKlZEnVFmGgoOc0xPGN6CNF7RDyCc7
-LJgaKBCWHx8PaSwOpcc+GBoAUXDh//tiUt5sD1vlc7vZKc49EHPZoUtuRbqHfNI9
-zCoKg+yO34A/0L4Syqehind6QVlfMcavtx4mgGA6olR9f6O/kxL9CKZYsjVje+SB
-3GFGL6rGhg5gxQKCAQEAgmPXwn/ExTmPaAZOIN9f3net4ransUzRzosutCTHk73D
-1CwihHEp21iNloPf9p4B+C8uUBojqx+gD/sZg0/xAr/F4ABkft5tanDDVCqcmwHd
-zbc8/tKvikMgJcArMAS3fQ1Joeeke3Rk3CkR7xLJX7V7lyIMfSa2rFUNEBQsTOoF
-QIRDsQ9WzOVUUld7g7fugvJI6B4H3DCtIES+umpmyg3MhfsBoFSEVCL7MZH73T5E
-zsYT5FUySQFPbBvHpPQ1tFzQOynddUm9YHgfFxG3iV/xBb/zoNEflnzmpH//7jKr
-yshMFvl/ayNGElHKo47UdPsu14ipHunLr++Ev5LOMQKCAQBJ05ZMkETlC8lAb/JY
-bKtb3SzeNSQpLAHq3LfledojvpR37aIt2AhOOjU1Uj3Ms4qV05NsjXpR1qdgdd6V
-ernaIP1MQSa/zfXx9v0yz0naLUWedTQbDdOddi1a5SVr2g8hrwBMwT/iK5IIfUjm
-TkDhG9yao3krbLctB2l9zLqKLyIXcZK6GY1YCRyS5T0G1JlOIGMR1BVXURdWlmRE
-3TGxDst8sshyyqXiGE2HlrpX89QwvAzSck+Yo1yTsFevtkyrD62DZc2kGx6bDBom
-rj/oqUdornyhS1agAn+Xx5ue8UexYCHiEyjInOR9PUa9FCYZJ2QlaW3H5gRbjAii
-pBzNAoIBAAP0k83hf8gbIvRbdJayhdRRjF7tK5ad7imp8/LhZqdYpkZG8CvxrGft
-R6xc+Hebl3XNJfTrLLyKY7dZt3MmSVKxiEg5+QAh49ih3nCG30A4Uu3Ime4GiKB6
-Twa7UdqSplWsLoXbUV4mPZQCzefUVdlDqYrO57eeSSmRrkJsdJONpjuhnehVVznp
-DhNNbvhGhsZarLLwBJ15SmQC0iOR6QVOFjJbgK5Ldpu3QBHDYpRqYMUEHBUkaohM
-HwKd3dtQHWJ4QS3VbF9CDs/tnlQH+xn9l/Qwl6WJcypxoKvATgOeJHrMCN4fQIll
-Lnmk2S0UPNTJwgZQsKgRGGsrF1v+Hlw=
------END PRIVATE KEY-----
