+++ internal/client/client.go
@@ -1,0 +1,29 @@
+package client
+
+import (
+	"github.com/pkg/errors"
+	"go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/credentials/insecure"
+)
+
+func NewClientConn(authority string, tls bool) (*grpc.ClientConn, error) {
+	var creds grpc.DialOption
+	if tls {
+		creds = grpc.WithTransportCredentials(credentials.NewTLS(nil))
+	} else {
+		creds = grpc.WithTransportCredentials(insecure.NewCredentials())
+	}
+
+	conn, err := grpc.Dial(authority,
+		creds,
+		grpc.WithUnaryInterceptor(otelgrpc.UnaryClientInterceptor()),
+		grpc.WithStreamInterceptor(otelgrpc.StreamClientInterceptor()),
+	)
+	if err != nil {
+		return nil, errors.Wrap(err, "dial tinkerbell server")
+	}
+
+	return conn, nil
+}
