+++ internal/e2e/testdata/01/hardware.yaml
