+++ internal/e2e/testdata/01/template.yaml
