+++ internal/e2e/testdata/02/hardware1.yaml
