+++ internal/e2e/testdata/02/template1.yaml
