+++ internal/e2e/testdata/02/template2.yaml
