+++ internal/e2e/testdata/02/template3.yaml
