--- internal/tests/errors.go
@@ -1,19 +1,0 @@
-package tests
-
-import "testing"
-
-// CompareErrors is a helper function for comparing an error value and a desired error.
-func CompareErrors(t *testing.T, got, want error) {
-	t.Helper()
-	if got != nil {
-		if want == nil {
-			t.Fatalf(`Got unexpected error: %v"`, got)
-		} else if got.Error() != want.Error() {
-			t.Fatalf(`Got unexpected error: got "%v" wanted "%v"`, got, want)
-		}
-		return
-	}
-	if got == nil && want != nil {
-		t.Fatalf("Missing expected error: %v", want)
-	}
-}
