+++ internal/testtime/frozen_time.go
@@ -1,4 +1,4 @@
-package tests
+package testtime
 
 import (
 	"time"
