+++ internal/testtime/frozen_time_test.go
@@ -1,4 +1,4 @@
-package tests
+package testtime
 
 import (
 	"testing"
